+++ CHANGELOG.md
@@ -17,6 +17,13 @@ Semantic versioning in our case means:
   change the client facing API, change code conventions signigicantly, etc.
   
   
+# 0.16.0
+
+### Features
+
+- Forbids the reassignment of variables that affect the generator expression
+
+
 ## 0.15.3 WIP
 
 ### Bugfixes
@@ -109,7 +116,6 @@ Semantic versioning in our case means:
 - Forbids inconsistent structuring of multiline comprehensions
 - Forbids to use unpythonic getters and setters such as `get_attribute` or `set_attribute`
 - Now `credits`, `license`, and `copyright` builtins are free to shadow
-- Forbids the reassignment of variables that affect the generator expression
 
 ### Bugfixes